Split filesystem handling into its own module
